ratatui_macros

Macro line

macro_rules! line {
    () => { ... };
    ($span:expr; $n:expr) => { ... };
    ($($span:expr),+ $(,)?) => { ... };
}

A macro for creating a Line using vec! syntax.

line! is similar to the vec! macro, but it returns a Line instead of a Vec.

Examples

Create a Line containing a vector of Spans:

use ratatui_macros::line;

let line = line!["hello", "world"];
let line = line!["hello".red(), "world".red().bold()];

Create a Line from a given Span repeated some amount of times:

let line = line!["hello"; 2];

Use span! macro inside line! macro for formatting.

use ratatui_macros::{line, span};

let line = line![span!("hello {}", "world"), span!(Modifier::BOLD; "goodbye {}", "world")];